gestion du slug

+++ Boom/Model/Model.php
@@ -18,22 +18,29 @@ class Model extends Table
         foreach ($_FILES as $key => $file) {
             if (empty($file['name'])) {
                 //$data->{$key} = null; yoloooo
-                $delete = !!$data[$key.'_delete'];
-                if ($delete) {
-                    foreach ($tabExt as $ext) {
-                        $filename = TARGET . $this->table() . '_' . $key. '.' . $ext;
-                        if (file_exists($filename)) {
-                            unlink($filename);
-                            $data[$key] = '';
-                            break;
+                if (isset($data[$key.'_delete'])) {
+                    $delete = !!$data[$key . '_delete'];
+                    if ($delete) {
+                        foreach ($tabExt as $ext) {
+                            $filename = TARGET . $this->table() . '_' . $key . '.' . $ext;
+                            if (file_exists($filename)) {
+                                unlink($filename);
+                                $data[$key] = '';
+                                break;
+                            }
                         }
                     }
                 }
             } else {
                 $extension  = pathinfo($_FILES[$key]['name'], PATHINFO_EXTENSION);
-                $nomImage = $this->table() . '_' . $key .'.'. $extension;
-                move_uploaded_file($_FILES[$key]['tmp_name'], TARGET.$nomImage);
-                $data[$key] = BASE_URL.TARGET.$nomImage;
+
+                if (in_array($extension, $tabExt)) {
+                    $nomImage = $this->table() . '_' . $key .'.'. $extension;
+                    move_uploaded_file($_FILES[$key]['tmp_name'], TARGET.$nomImage);
+                    $data[$key] = BASE_URL.TARGET.$nomImage;
+                } else {
+                    error(__('Le fichier ne correspond pas a une extention autorisée : ') . $extension);
+                }
             }
         }
     }
